ja: Introduce "Installation Guide" section
This patch separates "building environment procedure" from "Switching
Hub" section and adds "Installation Guide" section.

Also, this patch obsoletes the OpenFlow Tutorial VM image, because
the Ubuntu OS on this VM image is already EOL.
Instead of this VM image, this patch introduces the procedure for the
Ryu-Book Docker image.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>